edited transformer so legal types don't require sensitive data block
