+++ src/main/java/uk/gov/companieshouse/pscdataapi/util/PscTransformationHelper.java
@@ -17,21 +17,6 @@ public class PscTransformationHelper {
     }
 
     /**
-     * Creates the created and updated fields.
-     * @param requestBody request payload.
-     * @param pscDocument output document.
-     */
-    public static void createDateFields(FullRecordCompanyPSCApi requestBody,
-                                        PscDocument pscDocument) {
-        Created created = new Created();
-        created.setAt(LocalDateTime.parse(requestBody.getInternalData().getCreatedAt()));
-        Updated updated = new Updated();
-        updated.setAt(requestBody.getInternalData().getUpdatedAt());
-        pscDocument.setUpdated(updated);
-        pscDocument.setCreated(created);
-    }
-
-    /**
      * Creates Links field.
      * @param requestBody request payload.
      * @return Links object.
